+++ commercetools/commercetools-sdk-java-api/src/main/java-generated/com/commercetools/api/models/cart_discount/CartDiscountDraftBuilder.java
@@ -68,6 +68,9 @@ public class CartDiscountDraftBuilder implements Builder<CartDiscountDraft> {
     @Nullable
     private com.commercetools.api.models.discount_group.DiscountGroupResourceIdentifier discountGroup;
 
+    @Nullable
+    private com.commercetools.api.models.recurring_order.RecurringOrderScopeDraft recurringOrderScope;
+
     /**
      *  <p>Name of the CartDiscount.</p>
      * @param builder function to build the name value
@@ -480,6 +483,34 @@ public class CartDiscountDraftBuilder implements Builder<CartDiscountDraft> {
     }
 
     /**
+     *  <p>Scope of the Cart Discount for Recurring Orders.</p>
+     *  <p>If not set, the default is <a href="https://docs.commercetools.com/apis/ctp:api:type:NonRecurringOrdersOnlyDraft" rel="nofollow">NonRecurringOrdersOnlyDraft</a>.</p>
+     * @param recurringOrderScope value to be set
+     * @return Builder
+     */
+
+    public CartDiscountDraftBuilder recurringOrderScope(
+            @Nullable final com.commercetools.api.models.recurring_order.RecurringOrderScopeDraft recurringOrderScope) {
+        this.recurringOrderScope = recurringOrderScope;
+        return this;
+    }
+
+    /**
+     *  <p>Scope of the Cart Discount for Recurring Orders.</p>
+     *  <p>If not set, the default is <a href="https://docs.commercetools.com/apis/ctp:api:type:NonRecurringOrdersOnlyDraft" rel="nofollow">NonRecurringOrdersOnlyDraft</a>.</p>
+     * @param builder function to build the recurringOrderScope value
+     * @return Builder
+     */
+
+    public CartDiscountDraftBuilder recurringOrderScope(
+            Function<com.commercetools.api.models.recurring_order.RecurringOrderScopeDraftBuilder, Builder<? extends com.commercetools.api.models.recurring_order.RecurringOrderScopeDraft>> builder) {
+        this.recurringOrderScope = builder
+                .apply(com.commercetools.api.models.recurring_order.RecurringOrderScopeDraftBuilder.of())
+                .build();
+        return this;
+    }
+
+    /**
      *  <p>Name of the CartDiscount.</p>
      * @return name
      */
@@ -635,6 +666,17 @@ public class CartDiscountDraftBuilder implements Builder<CartDiscountDraft> {
     }
 
     /**
+     *  <p>Scope of the Cart Discount for Recurring Orders.</p>
+     *  <p>If not set, the default is <a href="https://docs.commercetools.com/apis/ctp:api:type:NonRecurringOrdersOnlyDraft" rel="nofollow">NonRecurringOrdersOnlyDraft</a>.</p>
+     * @return recurringOrderScope
+     */
+
+    @Nullable
+    public com.commercetools.api.models.recurring_order.RecurringOrderScopeDraft getRecurringOrderScope() {
+        return this.recurringOrderScope;
+    }
+
+    /**
      * builds CartDiscountDraft with checking for non-null required values
      * @return CartDiscountDraft
      */
@@ -643,7 +685,8 @@ public class CartDiscountDraftBuilder implements Builder<CartDiscountDraft> {
         Objects.requireNonNull(value, CartDiscountDraft.class + ": value is missing");
         Objects.requireNonNull(cartPredicate, CartDiscountDraft.class + ": cartPredicate is missing");
         return new CartDiscountDraftImpl(name, key, description, value, cartPredicate, target, sortOrder, stores,
-            isActive, validFrom, validUntil, requiresDiscountCode, stackingMode, custom, discountGroup);
+            isActive, validFrom, validUntil, requiresDiscountCode, stackingMode, custom, discountGroup,
+            recurringOrderScope);
     }
 
     /**
@@ -652,7 +695,8 @@ public class CartDiscountDraftBuilder implements Builder<CartDiscountDraft> {
      */
     public CartDiscountDraft buildUnchecked() {
         return new CartDiscountDraftImpl(name, key, description, value, cartPredicate, target, sortOrder, stores,
-            isActive, validFrom, validUntil, requiresDiscountCode, stackingMode, custom, discountGroup);
+            isActive, validFrom, validUntil, requiresDiscountCode, stackingMode, custom, discountGroup,
+            recurringOrderScope);
     }
 
     /**
@@ -685,6 +729,7 @@ public class CartDiscountDraftBuilder implements Builder<CartDiscountDraft> {
         builder.stackingMode = template.getStackingMode();
         builder.custom = template.getCustom();
         builder.discountGroup = template.getDiscountGroup();
+        builder.recurringOrderScope = template.getRecurringOrderScope();
         return builder;
     }