+++ commercetools/commercetools-sdk-java-api/src/main/java-generated/com/commercetools/api/models/recurring_order/NonRecurringOrdersOnlyImpl.java
@@ -1,0 +1,69 @@
+
+package com.commercetools.api.models.recurring_order;
+
+import java.time.*;
+import java.util.*;
+
+import com.fasterxml.jackson.annotation.JsonCreator;
+import com.fasterxml.jackson.databind.annotation.*;
+
+import io.vrap.rmf.base.client.ModelBase;
+import io.vrap.rmf.base.client.utils.Generated;
+
+import org.apache.commons.lang3.builder.EqualsBuilder;
+import org.apache.commons.lang3.builder.HashCodeBuilder;
+import org.apache.commons.lang3.builder.ToStringBuilder;
+import org.apache.commons.lang3.builder.ToStringStyle;
+
+/**
+ *  <p>Cart Discounts are applied to non-recurring Orders.</p>
+ */
+@Generated(value = "io.vrap.rmf.codegen.rendering.CoreCodeGenerator", comments = "https://github.com/commercetools/rmf-codegen")
+public class NonRecurringOrdersOnlyImpl implements NonRecurringOrdersOnly, ModelBase {
+
+    private String type;
+
+    /**
+     * create instance with all properties
+     */
+    @JsonCreator
+    public NonRecurringOrdersOnlyImpl() {
+        this.type = NON_RECURRING_ORDERS_ONLY;
+    }
+
+    /**
+     *
+     */
+
+    public String getType() {
+        return this.type;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o)
+            return true;
+
+        if (o == null || getClass() != o.getClass())
+            return false;
+
+        NonRecurringOrdersOnlyImpl that = (NonRecurringOrdersOnlyImpl) o;
+
+        return new EqualsBuilder().append(type, that.type).append(type, that.type).isEquals();
+    }
+
+    @Override
+    public int hashCode() {
+        return new HashCodeBuilder(17, 37).append(type).toHashCode();
+    }
+
+    @Override
+    public String toString() {
+        return new ToStringBuilder(this, ToStringStyle.SHORT_PREFIX_STYLE).append("type", type).build();
+    }
+
+    @Override
+    public NonRecurringOrdersOnly copyDeep() {
+        return NonRecurringOrdersOnly.deepCopy(this);
+    }
+}